Changed offsets from int to long. Changed punctuation removal.

+++ src/main/java/it/unipi/models/SkipPointerEntry.java
@@ -1,5 +1,5 @@
 package it.unipi.models;
 
-public record SkipPointerEntry(int docIdOffset, int freqOffset) {
+public record SkipPointerEntry(long docIdOffset, long freqOffset) {
 
 }
